+++ src/createproofdialog.cpp
@@ -11,14 +11,15 @@
 namespace WalletGUI
 {
 
-CreateProofDialog::CreateProofDialog(const QString& txHash, QWidget *parent)
+CreateProofDialog::CreateProofDialog(const QString& txHash, const QStringList& addresses, QWidget *parent)
     : QDialog(parent)
     , ui(new Ui::CreateProofDialog)
     , txHash_(txHash)
 {
     ui->setupUi(this);
-    showElements(false);
-    adjustSize();
+    ui->addressesBox->addItems(addresses);
+//    showElements(false);
+//    adjustSize();
 //    ui->verticalLayout->update();
 }
 
@@ -27,51 +28,51 @@ CreateProofDialog::~CreateProofDialog()
     delete ui;
 }
 
-void CreateProofDialog::showElements(bool show)
-{
-    ui->addressLabel->setVisible(show);
-    ui->addressesBox->setVisible(show);
-    ui->proofLabel->setVisible(show);
-    ui->proofEdit->setVisible(show);
-    ui->copyButton->setVisible(show);
-}
+//void CreateProofDialog::showElements(bool show)
+//{
+////    ui->addressLabel->setVisible(show);
+////    ui->addressesBox->setVisible(show);
+//    ui->proofLabel->setVisible(show);
+//    ui->proofEdit->setVisible(show);
+//    ui->copyButton->setVisible(show);
+//}
 
-QString CreateProofDialog::extractAddress(const QString& proof)
-{
-    static constexpr char ADDRESS_KEY[] = "address";
+//QString CreateProofDialog::extractAddress(const QString& proof)
+//{
+//    static constexpr char ADDRESS_KEY[] = "address";
 
-    QJsonParseError parseError;
-    QJsonDocument jsonDocument = QJsonDocument::fromJson(proof.toLatin1(), &parseError);
-    if (parseError.error != QJsonParseError::NoError)
-    {
-        ui->proofEdit->setTextColor(Qt::red);
-        ui->proofEdit->setText(parseError.errorString());
-        return QString{};
-    }
+//    QJsonParseError parseError;
+//    QJsonDocument jsonDocument = QJsonDocument::fromJson(proof.toLatin1(), &parseError);
+//    if (parseError.error != QJsonParseError::NoError)
+//    {
+//        ui->proofEdit->setTextColor(Qt::red);
+//        ui->proofEdit->setText(parseError.errorString());
+//        return QString{};
+//    }
 
-    if (!jsonDocument.isObject())
-    {
-        ui->proofEdit->setTextColor(Qt::red);
-        ui->proofEdit->setText(tr("JSON document is not an object."));
-        return QString{};
-    }
-    const QJsonObject json = jsonDocument.object();
-    const bool containsAddress = json.contains(ADDRESS_KEY);
-    if (!containsAddress)
-    {
-        ui->proofEdit->setTextColor(Qt::red);
-        ui->proofEdit->setText(tr("The proof does not contain any address."));
-        return QString{};
-    }
-    return json[ADDRESS_KEY].toString();
-}
+//    if (!jsonDocument.isObject())
+//    {
+//        ui->proofEdit->setTextColor(Qt::red);
+//        ui->proofEdit->setText(tr("JSON document is not an object."));
+//        return QString{};
+//    }
+//    const QJsonObject json = jsonDocument.object();
+//    const bool containsAddress = json.contains(ADDRESS_KEY);
+//    if (!containsAddress)
+//    {
+//        ui->proofEdit->setTextColor(Qt::red);
+//        ui->proofEdit->setText(tr("The proof does not contain any address."));
+//        return QString{};
+//    }
+//    return json[ADDRESS_KEY].toString();
+//}
 
 void CreateProofDialog::addProofs(const QStringList& proofs)
 {
     ui->proofEdit->clear();
-    ui->addressesBox->clear();
+//    ui->addressesBox->clear();
     proofs_.clear();
-    showElements(true);
+//    showElements(true);
 
     if (proofs.isEmpty())
     {
@@ -80,20 +81,29 @@ void CreateProofDialog::addProofs(const QStringList& proofs)
         return;
     }
 
-    for (int i = 0; i < proofs.size(); ++i)
+    if (proofs.size() > 1)
     {
-        const QString address = extractAddress(proofs[i]);
-        if (address.isEmpty())
-            continue;
-        proofs_.insert(i, proofs[i]);
-        ui->addressesBox->addItem(address);
+        ui->proofEdit->setTextColor(Qt::red);
+        ui->proofEdit->setText(tr("Several proofs were generated."));
+        return;
     }
-//    ui->proofEdit->setText(proof);
+
+//    for (int i = 0; i < proofs.size(); ++i)
+//    {
+//        const QString address = extractAddress(proofs[i]);
+//        if (address.isEmpty())
+//            continue;
+//        proofs_.insert(i, proofs[i]);
+//        ui->addressesBox->addItem(address);
+//    }
+    const int index = ui->addressesBox->currentIndex();
+    proofs_.insert(index, proofs.front());
+    ui->proofEdit->setText(proofs.front());
 }
 
 void CreateProofDialog::generateProof()
 {
-    emit generateProofSignal(txHash_, ui->messageEdit->text());
+    emit generateProofSignal(txHash_, ui->addressesBox->currentText(), ui->messageEdit->text());
 }
 
 void CreateProofDialog::copyToClipboard()